Refactor Region into RegionConverter and add tests
